Update WebhookController.php
The issue is that you cannot test Stripe Webhooks because of the brute force protection.  The test event from Stripe.com has an id of evt_00000000000000 which obviously cannot be validated.  This makes it impossible to test.  

So if in debug mode allow this protection to be bypassed.  Or maybe add in a services.stripe.testing or something like that.