Allow testing of Webhooks using ENV
After seeing the pull request #190 I wanted to contribute with what looks to me a better solution.

As @taylorotwell said using environment variables looked like a better choice.

To keep it simple I added a isInTestingMode method that return a boolean according to the CASHIER_TESTING env variable.

Then in the handleWebhook method we can verify that we are not in testing mode.