replaceCard() method rather than updateCard()
May I suggest you to change the way the `updateCard()` method works or perhaps add a new method called `replaceCard()` that replaces the existing card?

**The problem with the current `updateCard` method:**

With the `updateCard()` method, the card is added to the sources list and then sets the new card as the `default_source`. the result of this method has 2 outcomes:

1. Multiple cards gets added to the list although the recent one is set as `default_source`

2. When updating the card using this method, if there are any unpaid invoices (i.e. invoices in `past_due` state), they are not getting automatically charged during the card update.

In order for stripe to re-attempt charging customer on all invoices in `past_due` state, the `source` parameter needs to be passed. So I have created a new method called `replaceCard()`. 

What do you think? Can this be added to Cashier?