Fix: PDF Options, Allow General PDF Format Configuration
The code allowing customisation of the PDF invoices does not match what
is available in the version of dompdf required by cashier.  These
changes allow users to optionally use Laravel style configuration file
that can be shared with any part of the application that uses the
library to generate PDF files.