Fix customer not set when present
When a single charge is performed and the billable entity is a stripe customer it's not set when a custom source is passed. This PR fixes this behavior to allow single charges to stripe customers with a different payment source. If the payment source isn't attached to the customer the Stripe API call will fail. This is now also covered with tests.