Implement config file
Cashier is at the moment one of the only Laravel packages which doesn't uses a config file. I think it would be better if it was more in line of the other packages by making use of one. This is what people are used to as well. These changes add a config file and cleans up the way configuration is handled within cashier.

Another small change is the rename of STRIPE_MODEL to CASHIER_MODEL since that env variable is Cashier specfic and not Stripe specific.

This also allows us to easily add new config options in later pull requests (like the email reminder option in the payment intents PR).

Closes https://github.com/laravel/cashier/issues/531