Cover more downgrade scenarios
When a plan downgrade fails because of a next action requirement for the payment or a faulty card, we'll not put the subscription into an incomplete state because it could be that the person is upgrading to a yearly plan with a prorated invoice. This would mean that the customer still paid for the current period and would be allowed to continue on their subscription. These scenarios are now covered by these tests.