Fix findInvoiceOrFail behavior
Because of the changes made to the constructor of the Invoice object recently, the subsequent findInvoiceOrFail method didn't throw the AccessDeniedHttpException anymore and instead threw the new Exception from the constructor. I've now fixed this method to properly throw the AccessDeniedHttpException again.

I've also changed the Exception classes to two dedicated Exception classes which people can type-hint against.