Remove unique index from items table
I stumbled upon an edge case where Stripe created a new subscription item for the same price ID. This means Cashier attempts to insert this while the unique constraint is active. Although it's unlikely this happens often (we only saw this just a couple of times) we shouldn't rely on this index anymore.